requirements.txt
torch
transformers
Pillow
requests
opencv-python
numpy
plotly
fastapi
uvicorn
matplotlib